requirements.txt
streamlit
pandas
numpy
plotly
psycopg2-binary
python-dotenv
pyarrow
